autoscaler/requirements.txt
requests==2.28.2
google-auth==2.16.2
google-cloud-monitoring==2.14.1
google-cloud-logging==3.5.0